De-dup static rendering
